refactor: move Ethereum client and configuration into separate modules
