+++ chain-signatures/node/src/indexer_eth/config.rs
@@ -1,0 +1,42 @@
+use std::fmt;
+
+#[derive(Clone)]
+pub struct EthConfig {
+    /// The ethereum account secret key used to sign eth respond txn.
+    pub account_sk: String,
+    /// Ethereum consensus HTTP RPC URL
+    pub consensus_rpc_http_url: String,
+    /// Ethereum execution HTTP RPC URL
+    pub execution_rpc_http_url: String,
+    /// The contract address to watch without the `0x` prefix
+    pub contract_address: String,
+    /// must be one of sepolia, mainnet
+    pub network: String,
+    /// path to store helios data
+    pub helios_data_path: String,
+    /// refresh finalized block interval in milliseconds
+    pub refresh_finalized_interval: u64,
+    /// Enable the indexer to just send requests optimistically instead waiting for final.
+    pub optimistic_requests: bool,
+    /// light client is true if using helios, false if using direct rpc
+    pub light_client: bool,
+}
+
+impl fmt::Debug for EthConfig {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.debug_struct("EthConfig")
+            .field("account_sk", &"<hidden>")
+            .field("consensus_rpc_http_url", &self.consensus_rpc_http_url)
+            .field("execution_rpc_http_url", &self.execution_rpc_http_url)
+            .field("contract_address", &self.contract_address)
+            .field("network", &self.network)
+            .field("helios_data_path", &self.helios_data_path)
+            .field(
+                "refresh_finalized_interval",
+                &self.refresh_finalized_interval,
+            )
+            .field("optimistic_requests", &self.optimistic_requests)
+            .field("light_client", &self.light_client)
+            .finish()
+    }
+}
