+++ chain-signatures/node/src/indexer_eth/mod.rs
@@ -1,4 +1,6 @@
 pub mod abi;
+mod client;
+mod config;
 pub mod indexer_eth_direct_rpc;
 #[cfg(feature = "helios")]
 pub mod indexer_eth_helios;
@@ -8,7 +10,6 @@ use crate::metrics::requests::{record_request_latency_since, SignRequestStep};
 use crate::protocol::Chain;
 use crate::respond_bidirectional::CompletedTx;
 use crate::stream::{ChainIndexer, ChainStream};
-use crate::util::retry;
 
 use alloy::eips::BlockNumberOrTag;
 use alloy::primitives::hex::{self, ToHexExt};
@@ -17,6 +18,8 @@ use alloy::rpc::types::{Block, BlockId, Log};
 use alloy::sol_types::SolEvent;
 use anyhow::Context as _;
 use async_trait::async_trait;
+pub use client::EthereumClient;
+pub use config::EthConfig;
 use futures_util::stream;
 use k256::elliptic_curve::sec1::FromEncodedPoint;
 use k256::{AffinePoint as K256AffinePoint, EncodedPoint, FieldBytes, Scalar};
@@ -26,10 +29,8 @@ use mpc_primitives::{
     SignId, Signature as MpcSignature, SignatureRespondedEvent, StateManager,
     LATEST_MPC_KEY_VERSION, MAX_SECP256K1_SCALAR,
 };
-use secrecy::{ExposeSecret, SecretString};
 use serde::{Deserialize, Serialize};
 use std::collections::{HashMap, HashSet};
-use std::fmt;
 use std::str::FromStr;
 use std::sync::Arc;
 use tokio::sync::{mpsc, Notify};
@@ -122,217 +123,6 @@ impl BlockAndRequests {
             indexed_requests,
             respond_logs,
             execution_events,
-        }
-    }
-}
-
-#[derive(Clone)]
-pub struct EthConfig {
-    /// The ethereum account secret key used to sign eth respond txn.
-    pub account_sk: String,
-    /// Ethereum consensus HTTP RPC URL
-    pub consensus_rpc_http_url: String,
-    /// Ethereum execution HTTP RPC URL
-    pub execution_rpc_http_url: String,
-    /// The contract address to watch without the `0x` prefix
-    pub contract_address: String,
-    /// must be one of sepolia, mainnet
-    pub network: String,
-    /// path to store helios data
-    pub helios_data_path: String,
-    /// refresh finalized block interval in milliseconds
-    pub refresh_finalized_interval: u64,
-    /// Enable the indexer to just send requests optimistically instead waiting for final.
-    pub optimistic_requests: bool,
-    /// light client is true if using helios, false if using direct rpc
-    pub light_client: bool,
-}
-
-impl fmt::Debug for EthConfig {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_struct("EthConfig")
-            .field("account_sk", &"<hidden>")
-            .field("consensus_rpc_http_url", &self.consensus_rpc_http_url)
-            .field("execution_rpc_http_url", &self.execution_rpc_http_url)
-            .field("contract_address", &self.contract_address)
-            .field("network", &self.network)
-            .field("helios_data_path", &self.helios_data_path)
-            .field(
-                "refresh_finalized_interval",
-                &self.refresh_finalized_interval,
-            )
-            .field("optimistic_requests", &self.optimistic_requests)
-            .field("light_client", &self.light_client)
-            .finish()
-    }
-}
-
-// Configures Ethereum indexer.
-#[derive(Debug, Clone, clap::Parser)]
-#[group(id = "indexer_eth_options")]
-pub struct EthArgs {
-    // -- Core --
-    /// The ethereum account secret key used to sign eth respond txn.
-    #[arg(
-        long,
-        env("MPC_ETH_ACCOUNT_SK"),
-        requires_all = ["eth_execution_rpc_http_url", "eth_contract_address"]
-    )]
-    pub eth_account_sk: Option<SecretString>,
-    /// The contract address to watch without the `0x` prefix
-    #[clap(long, env("MPC_ETH_CONTRACT_ADDRESS"), requires = "eth_account_sk")]
-    pub eth_contract_address: Option<String>,
-
-    // -- RPC endpoints --
-    /// Ethereum execution RPC URL
-    #[clap(
-        long,
-        env("MPC_ETH_EXECUTION_RPC_HTTP_URL"),
-        requires = "eth_account_sk"
-    )]
-    pub eth_execution_rpc_http_url: Option<String>,
-
-    // -- Helios light-client --
-    /// Use Helios light client instead of direct RPC
-    #[clap(
-        long,
-        env("MPC_ETH_LIGHT_CLIENT"),
-        default_value = "false",
-        requires_if("true", "eth_consensus_rpc_http_url")
-    )]
-    pub eth_light_client: bool,
-    /// Ethereum consensus RPC URL (required when --eth-light-client is set)
-    #[clap(
-        long,
-        env("MPC_ETH_CONSENSUS_RPC_HTTP_URL"),
-        requires = "eth_account_sk"
-    )]
-    pub eth_consensus_rpc_http_url: Option<String>,
-    /// The network that the eth indexer is running on. Either "sepolia"/"mainnet"
-    #[clap(
-        long,
-        env("MPC_ETH_NETWORK"),
-        requires = "eth_account_sk",
-        default_value = "sepolia",
-        value_parser = ["sepolia", "mainnet"],
-    )]
-    pub eth_network: Option<String>,
-    /// Helios light client data path
-    #[clap(
-        long,
-        env("MPC_ETH_HELIOS_DATA_PATH"),
-        requires = "eth_account_sk",
-        default_value = "/helios/sepolia"
-    )]
-    pub eth_helios_data_path: Option<String>,
-
-    // -- Behaviour --
-    /// Refresh finalized block interval in milliseconds
-    #[clap(
-        long,
-        env("MPC_ETH_REFRESH_FINALIZED_INTERVAL"),
-        default_value = "10000"
-    )]
-    pub eth_refresh_finalized_interval: Option<u64>,
-    /// Enable the indexer to just send requests optimistically instead waiting for final.
-    /// Useful for testing where we do not want to reach finality due to how long it takes.
-    #[clap(long, env("MPC_ETH_OPTIMISTIC_REQUESTS"), default_value = "false")]
-    pub eth_optimistic_requests: bool,
-}
-
-impl EthArgs {
-    pub fn into_str_args(self) -> Vec<String> {
-        let mut args = Vec::with_capacity(10);
-        if let Some(eth_account_sk) = self.eth_account_sk {
-            args.extend([
-                "--eth-account-sk".to_string(),
-                eth_account_sk.expose_secret().to_string(),
-            ]);
-        }
-        if let Some(eth_consensus_rpc_http_url) = self.eth_consensus_rpc_http_url {
-            args.extend([
-                "--eth-consensus-rpc-http-url".to_string(),
-                eth_consensus_rpc_http_url,
-            ]);
-        }
-        if let Some(eth_execution_rpc_http_url) = self.eth_execution_rpc_http_url {
-            args.extend([
-                "--eth-execution-rpc-http-url".to_string(),
-                eth_execution_rpc_http_url,
-            ]);
-        }
-        if let Some(eth_contract_address) = self.eth_contract_address {
-            args.extend(["--eth-contract-address".to_string(), eth_contract_address]);
-        }
-        if let Some(eth_network) = self.eth_network {
-            args.extend(["--eth-network".to_string(), eth_network]);
-        }
-        if let Some(eth_helios_data_path) = self.eth_helios_data_path {
-            args.extend(["--eth-helios-data-path".to_string(), eth_helios_data_path]);
-        }
-        if let Some(eth_refresh_finalized_interval) = self.eth_refresh_finalized_interval {
-            args.extend([
-                "--eth-refresh-finalized-interval".to_string(),
-                eth_refresh_finalized_interval.to_string(),
-            ]);
-        }
-        if self.eth_optimistic_requests {
-            args.push("--eth-optimistic-requests".to_string());
-        }
-        if self.eth_light_client {
-            args.push("--eth-light-client".to_string());
-        }
-        args
-    }
-
-    pub fn into_config(self) -> Option<EthConfig> {
-        #[cfg(not(feature = "helios"))]
-        if self.eth_light_client {
-            tracing::warn!(
-                "ignoring ethereum light client request because mpc-node was built without helios feature"
-            );
-        }
-
-        Some(EthConfig {
-            account_sk: self.eth_account_sk?.expose_secret().to_string(), // this is safe because  EthConfig has custom Debug implementation that redacts the account_sk field
-            consensus_rpc_http_url: self.eth_consensus_rpc_http_url.unwrap_or_default(),
-            execution_rpc_http_url: self.eth_execution_rpc_http_url.unwrap(),
-            contract_address: self.eth_contract_address.unwrap(),
-            network: self.eth_network.unwrap_or_default(),
-            helios_data_path: self.eth_helios_data_path.unwrap_or_default(),
-            refresh_finalized_interval: self.eth_refresh_finalized_interval.unwrap(),
-            optimistic_requests: self.eth_optimistic_requests,
-            #[cfg(feature = "helios")]
-            light_client: self.eth_light_client,
-            #[cfg(not(feature = "helios"))]
-            light_client: false,
-        })
-    }
-
-    pub fn from_config(config: Option<EthConfig>) -> Self {
-        match config {
-            Some(config) if !config.account_sk.is_empty() => Self {
-                eth_account_sk: Some(config.account_sk.into()),
-                eth_consensus_rpc_http_url: Some(config.consensus_rpc_http_url),
-                eth_execution_rpc_http_url: Some(config.execution_rpc_http_url),
-                eth_contract_address: Some(config.contract_address),
-                eth_network: Some(config.network),
-                eth_helios_data_path: Some(config.helios_data_path),
-                eth_refresh_finalized_interval: Some(config.refresh_finalized_interval),
-                eth_optimistic_requests: config.optimistic_requests,
-                eth_light_client: config.light_client,
-            },
-            _ => Self {
-                eth_account_sk: None,
-                eth_consensus_rpc_http_url: None,
-                eth_execution_rpc_http_url: None,
-                eth_contract_address: None,
-                eth_network: None,
-                eth_helios_data_path: None,
-                eth_refresh_finalized_interval: None,
-                eth_optimistic_requests: false,
-                eth_light_client: false,
-            },
         }
     }
 }
@@ -567,272 +357,6 @@ impl SignatureRequestedEvent {
     pub fn generate_request_id(&self) -> [u8; 32] {
         let abi_encoded = self.encode_abi();
         alloy::primitives::keccak256(abi_encoded).into()
-    }
-}
-
-#[derive(Clone)]
-pub enum EthereumClient {
-    #[cfg(feature = "helios")]
-    Helios(indexer_eth_helios::HeliosEthereumClient),
-    DirectRpc(indexer_eth_direct_rpc::RpcEthereumClient),
-}
-
-impl EthereumClient {
-    pub async fn new(eth: EthConfig) -> anyhow::Result<EthereumClient> {
-        if eth.light_client {
-            #[cfg(feature = "helios")]
-            {
-                return Ok(EthereumClient::Helios(
-                    indexer_eth_helios::build_client(eth.clone()).await?,
-                ));
-            }
-
-            #[cfg(not(feature = "helios"))]
-            {
-                anyhow::bail!(
-                    "ethereum light client requested, but mpc-node was built without helios feature"
-                );
-            }
-        }
-
-        {
-            Ok(EthereumClient::DirectRpc(
-                indexer_eth_direct_rpc::RpcEthereumClient::new(&eth.execution_rpc_http_url),
-            ))
-        }
-    }
-
-    fn client_name(&self) -> &str {
-        match self {
-            #[cfg(feature = "helios")]
-            EthereumClient::Helios(_) => "Helios",
-            EthereumClient::DirectRpc(_) => "DirectRpc",
-        }
-    }
-
-    async fn get_block(&self, block_id: BlockId) -> Option<Block> {
-        // Configure retry behaviour and delegate to shared retry_async helper.
-        let retry_config = retry::RetryConfig::default();
-        let get_block_op = |_attempt: usize| async {
-            match self {
-                #[cfg(feature = "helios")]
-                EthereumClient::Helios(client) => client.get_block(block_id).await,
-                EthereumClient::DirectRpc(client) => client.get_block(block_id).await,
-            }
-        };
-
-        let res = retry::retry_async(
-            retry_config,
-            get_block_op,
-            |_attempt, _reason| true,
-            |attempt, reason, sleep_duration| match reason {
-                retry::RetryReason::Error(e) => {
-                    tracing::warn!(
-                        client = self.client_name(),
-                        "get_block failed (attempt {attempt}) for {block_id:?}: {e:#}; retrying in {sleep_duration:?}"
-                    );
-                }
-                retry::RetryReason::Timeout(t) => {
-                    tracing::warn!(
-                        client = self.client_name(),
-                        "get_block timed out after {t:?} (attempt {attempt}) for {block_id:?}; retrying in {sleep_duration:?}"
-                    );
-                }
-            },
-        )
-        .await;
-
-        match res {
-            Ok(Some(block)) => Some(block),
-            Ok(None) => {
-                tracing::warn!(client = self.client_name(), "Block {block_id:?} not found");
-                None
-            }
-            Err(retry::RetryError::Exhausted {
-                attempts,
-                last_error,
-            }) => {
-                tracing::warn!(
-                    client = self.client_name(),
-                    "get_block failed for {block_id:?}: {last_error:#}; exhausted after {attempts} attempts"
-                );
-                None
-            }
-            Err(retry::RetryError::TimeoutExhausted {
-                attempts,
-                last_timeout,
-            }) => {
-                tracing::warn!(
-                    client = self.client_name(),
-                    "get_block timed out for {block_id:?} (last timeout {last_timeout:?}); exhausted after {attempts} attempts"
-                );
-                None
-            }
-        }
-    }
-
-    async fn get_blocks(&self, block_ids: &[BlockId]) -> Vec<MaybeBlock> {
-        if block_ids.is_empty() {
-            return Vec::new();
-        }
-
-        let retry_config = retry::RetryConfig::default();
-        let block_ids = block_ids.to_vec();
-        let get_blocks_op = |_attempt: usize| {
-            let block_ids = block_ids.clone();
-            async move {
-                match self {
-                    #[cfg(feature = "helios")]
-                    EthereumClient::Helios(client) => client.get_blocks(&block_ids).await,
-                    EthereumClient::DirectRpc(client) => client.get_blocks(&block_ids).await,
-                }
-            }
-        };
-
-        match retry::retry_async(
-            retry_config,
-            get_blocks_op,
-            |_attempt, _reason| true,
-            |attempt, reason, sleep_duration| match reason {
-                retry::RetryReason::Error(e) => {
-                    tracing::warn!(
-                        client = self.client_name(),
-                        num_blocks = block_ids.len(),
-                        "get_blocks failed (attempt {attempt}): {e:#}; retrying in {sleep_duration:?}"
-                    );
-                }
-                retry::RetryReason::Timeout(t) => {
-                    tracing::warn!(
-                        client = self.client_name(),
-                        num_blocks = block_ids.len(),
-                        "get_blocks timed out after {t:?} (attempt {attempt}); retrying in {sleep_duration:?}"
-                    );
-                }
-            },
-        )
-        .await
-        {
-            Ok(blocks) => blocks,
-            Err(retry::RetryError::Exhausted { attempts, last_error }) => {
-                tracing::warn!(
-                    client = self.client_name(),
-                    num_blocks = block_ids.len(),
-                    "get_blocks failed: {last_error:#}; exhausted after {attempts} attempts"
-                );
-                block_ids.iter().copied().map(MaybeBlock::Missing).collect()
-            }
-            Err(retry::RetryError::TimeoutExhausted { attempts, last_timeout }) => {
-                tracing::warn!(
-                    client = self.client_name(),
-                    num_blocks = block_ids.len(),
-                    "get_blocks timed out (last timeout {last_timeout:?}); exhausted after {attempts} attempts"
-                );
-                block_ids.iter().copied().map(MaybeBlock::Missing).collect()
-            }
-        }
-    }
-
-    async fn get_block_receipts(
-        &self,
-        block_id: BlockId,
-    ) -> anyhow::Result<Option<Vec<alloy::rpc::types::TransactionReceipt>>> {
-        match self {
-            #[cfg(feature = "helios")]
-            EthereumClient::Helios(client) => client.get_block_receipts(block_id).await,
-            EthereumClient::DirectRpc(client) => client.get_block_receipts(block_id).await,
-        }
-    }
-
-    async fn get_nonce(&self, address: Address, block_id: BlockId) -> anyhow::Result<u64> {
-        match self {
-            #[cfg(feature = "helios")]
-            EthereumClient::Helios(client) => client.get_nonce(address, block_id).await,
-            EthereumClient::DirectRpc(client) => client.get_nonce(address, block_id).await,
-        }
-    }
-
-    pub fn block_number_from_id(block_id: BlockId) -> BlockNumber {
-        match block_id {
-            BlockId::Number(BlockNumberOrTag::Number(block_number)) => block_number,
-            BlockId::Number(tag) => panic!("expected numbered block id, got {tag:?}"),
-            BlockId::Hash(hash) => panic!("expected numbered block id, got hash {hash:?}"),
-        }
-    }
-
-    pub async fn get_transaction_by_hash(
-        &self,
-        tx_hash: alloy::primitives::B256,
-    ) -> anyhow::Result<Option<alloy::rpc::types::Transaction>> {
-        match self {
-            #[cfg(feature = "helios")]
-            EthereumClient::Helios(client) => client.get_transaction_by_hash(tx_hash).await,
-            EthereumClient::DirectRpc(client) => client.get_transaction_by_hash(tx_hash).await,
-        }
-    }
-
-    pub async fn trace_transaction_output(
-        &self,
-        tx_hash: alloy::primitives::B256,
-    ) -> anyhow::Result<alloy::primitives::Bytes> {
-        match self {
-            #[cfg(feature = "helios")]
-            EthereumClient::Helios(client) => client.trace_transaction_output(tx_hash).await,
-            EthereumClient::DirectRpc(client) => client.trace_transaction_output(tx_hash).await,
-        }
-    }
-
-    pub async fn call(
-        &self,
-        from: Address,
-        to: Address,
-        data: Bytes,
-        block_number: u64,
-    ) -> anyhow::Result<Bytes> {
-        match self {
-            #[cfg(feature = "helios")]
-            EthereumClient::Helios(client) => client.call(from, to, data, block_number).await,
-            EthereumClient::DirectRpc(client) => client.call(from, to, data, block_number).await,
-        }
-    }
-
-    async fn get_latest_block_number(&self) -> Option<u64> {
-        self.get_block(BlockId::Number(alloy::rpc::types::BlockNumberOrTag::Latest))
-            .await
-            .map(|block| block.header.number)
-    }
-
-    fn clamp_oldest_supported(
-        &self,
-        requested_start: u64,
-        anchor_height: BlockNumber,
-    ) -> BlockNumber {
-        let max_catchup_blocks = match self {
-            #[cfg(feature = "helios")]
-            EthereumClient::Helios(_) => indexer_eth_helios::MAX_CATCHUP_BLOCKS,
-            EthereumClient::DirectRpc(_) => indexer_eth_direct_rpc::MAX_CATCHUP_BLOCKS,
-        };
-        Self::clamp_oldest_supported_with(requested_start, anchor_height, max_catchup_blocks)
-    }
-
-    fn clamp_oldest_supported_with(
-        requested_start: u64,
-        anchor_height: BlockNumber,
-        max_catchup_blocks: u64,
-    ) -> BlockNumber {
-        let catchup_end = anchor_height.saturating_sub(1);
-        let oldest_supported = catchup_end.saturating_sub(max_catchup_blocks);
-
-        if requested_start < oldest_supported {
-            tracing::warn!(
-                requested_start,
-                anchor_height,
-                oldest_supported,
-                "ethereum catchup start is older than supported range; clamping"
-            );
-            oldest_supported
-        } else {
-            requested_start
-        }
     }
 }
